add signup with error message

+++ client/src/actions/index.js
@@ -18,12 +18,16 @@ export function signupUser({ email, password }) {
   return function(dispatch) {
     axios.post(`${ROOT_URL}/signup`, { email, password })
       .then(response => {
-        dispatch({
-          type: AUTH_USER,
-          payload: email
-        });
-        localStorage.setItem('token', response.data.token);
-        browserHistory.push('/securepage');
+        if (response.data.message === 'User Already Exists') {
+            dispatch(authError(response.data.error))
+        } else {
+            dispatch({
+              type: AUTH_USER,
+              payload: email
+            });
+            localStorage.setItem('token', response.data.token);
+            browserHistory.push('/securepage');
+        }
       })
       .catch(response => dispatch(authError(response.data.error)));
   }
